AGREGANDO CATEGORIAS, PRODUCTOS Y BUSQUEDA

+++ server/routes/categoria.js
@@ -1,5 +1,5 @@
 const express = require('express');
-let { verificaToken } = require('../middlewares/autenticacion');
+let { verificaToken , verificaAdminRole } = require('../middlewares/autenticacion');
 
 let app = express();
 
@@ -10,18 +10,24 @@ let Category = require('../models/categoria');
 //=================================
 
 app.get("/categoria", verificaToken, (req, res) => {
-  Category.find({}, "description userId").exec((err, categoryDB) => {
+  Category.find({}, "description user")
+  .populate('user','name')
+  .exec((err, categoryDB) => {
     if (err) {
       return res
-        .status(400)
-        .json({ ok: false, err: { message: "No hay categorias" } });
+        .status(500)
+        .json({ ok: false, err });
     }
-    if (categoryDB) {
+    if (!categoryDB) {
+      return res
+        .status(400)
+        .json({ ok: false, err: { message: "No se encontro la categorias" } });
+    }
+ 
       res.json({
         ok: true,
         category: categoryDB
       });
-    }
   });
 });
 
@@ -35,7 +41,12 @@ app.get("/categoria/:id", verificaToken, (req, res) => {
     if (err) {
       return res
         .status(500)
-        .json({ ok: false, err: { message: "No se encomtro la categorias" } });
+        .json({ ok: false, err });
+    }
+    if (!categoryDB) {
+      return res
+        .status(400)
+        .json({ ok: false, err: { message: "No se encontro la categorias" } });
     }
     res.json({
       ok: true,
@@ -54,7 +65,7 @@ app.post("/categoria", verificaToken, (req, res)=>{
 
     let addCategory = new Category({
       description: category.description,
-      userId: id
+      user: id
     });
     addCategory.save((err, categoryDB) => {
       if (err) {
@@ -65,6 +76,7 @@ app.post("/categoria", verificaToken, (req, res)=>{
             err
           });
       }
+      console.log( categoryDB );
         if (!categoryDB) {
             return res
                 .status(400)
@@ -84,9 +96,8 @@ app.put("/categoria/:id", verificaToken, (req, res) =>{
     
     let id = req.params.id;
     let description = req.body.description;
-    console.log(id);
   // Categiria.findById(...);
-    Category.findOneAndUpdate({ _id:id }, { description }, { new: true, runValidators: true },(err,categoryDB)=>{
+  Category.findOneAndUpdate({ _id: id }, { description:description },{ new:true,runValidators:true },(err,categoryDB)=>{
         if (err) {
             return res.status(500).json({ ok: false, err:{message:'No se encontro el ID'} });
         }
@@ -105,14 +116,19 @@ app.put("/categoria/:id", verificaToken, (req, res) =>{
 app.delete("/categoria/:id", verificaToken, (req, res) => {
   let _id = req.params.id;
   // Categoria.findById(...);
-  Category.findOneAndDelete({ _id: _id }, (err, deleteCategory) => {
+  Category.findOneAndDelete({ _id: _id }, [verificaToken, verificaAdminRole], (err, deleteCategory) => {
     if (err) {
+      return res
+        .status(500)
+        .json({ ok: false, err: err });
+    }
+    if (!deleteCategory) {
       return res
         .status(400)
         .json({ ok: false, err: { message: "No se encontro el ID" } });
     }
 
-    res.json({ ok: true, category: deleteCategory });
+    res.json({ ok: true, message: "Categoria eliminada" });
   });
 });
